Rework zsh config & make install script idempotent

+++ .devcontainer/zsh/install-zsh.sh
@@ -2,9 +2,11 @@
 set -e
 
 # zsh
-sudo apt update \
-    && DEBIAN_FRONTEND=noninteractive sudo apt install -y \
-       -o Dpkg::Options::="--force-confnew" zsh
+if ! command -v zsh &> /dev/null; then
+    sudo apt update \
+        && DEBIAN_FRONTEND=noninteractive sudo apt install -y \
+           -o Dpkg::Options::="--force-confnew" zsh
+fi
 
 # zsh history
 # adapted from https://code.visualstudio.com/remote/advancedcontainers/persist-bash-history
@@ -17,29 +19,39 @@ cp .devcontainer/zsh/.zshrc ~/.zshrc
 
 # Powerline fonts
 # https://github.com/powerline/fonts?tab=readme-ov-file#quick-installation
-git clone --depth=1 https://github.com/powerline/fonts.git
-cd fonts
-./install.sh
-cd .. && rm -rf fonts
+if [ ! -d ~/.local/share/fonts ] || [ -z "$(find ~/.local/share/fonts -name '*Powerline*' 2>/dev/null)" ]; then
+    git clone --depth=1 https://github.com/powerline/fonts.git
+    cd fonts
+    ./install.sh
+    cd .. && rm -rf fonts
+fi
 
 # Powerlevel10k
 # https://github.com/romkatv/powerlevel10k?tab=readme-ov-file#manual
-git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ~/powerlevel10k
+if [ ! -d ~/powerlevel10k ]; then
+    git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ~/powerlevel10k
+fi
 echo 'source ~/powerlevel10k/powerlevel10k.zsh-theme' >> ~/.zshrc
 cp .devcontainer/zsh/.p10k.zsh ~/.p10k.zsh
 echo '[[ ! -f ~/.p10k.zsh ]] || source ~/.p10k.zsh' >> ~/.zshrc
 
 # zsh autosuggestion
 # https://github.com/zsh-users/zsh-autosuggestions/blob/master/INSTALL.md#manual-git-clone
-git clone --depth=1 https://github.com/zsh-users/zsh-autosuggestions ~/.zsh/zsh-autosuggestions
+if [ ! -d ~/.zsh/zsh-autosuggestions ]; then
+    git clone --depth=1 https://github.com/zsh-users/zsh-autosuggestions ~/.zsh/zsh-autosuggestions
+fi
 echo 'source ~/.zsh/zsh-autosuggestions/zsh-autosuggestions.zsh' >> ~/.zshrc
 
 # zsh syntax highlighting
 # https://github.com/zsh-users/zsh-syntax-highlighting/blob/master/INSTALL.md#in-your-zshrc
-git clone --depth=1 https://github.com/zsh-users/zsh-syntax-highlighting.git ~/.zsh/zsh-syntax-highlighting
+if [ ! -d ~/.zsh/zsh-syntax-highlighting ]; then
+    git clone --depth=1 https://github.com/zsh-users/zsh-syntax-highlighting.git ~/.zsh/zsh-syntax-highlighting
+fi
 echo "source ~/.zsh/zsh-syntax-highlighting/zsh-syntax-highlighting.zsh" >> ~/.zshrc
 
 # zsh history substring search (install AFTER syntax highlighting)
 # https://github.com/zsh-users/zsh-history-substring-search
-git clone --depth=1 https://github.com/zsh-users/zsh-history-substring-search.git ~/.zsh/zsh-history-substring-search
+if [ ! -d ~/.zsh/zsh-history-substring-search ]; then
+    git clone --depth=1 https://github.com/zsh-users/zsh-history-substring-search.git ~/.zsh/zsh-history-substring-search
+fi
 echo "source ~/.zsh/zsh-history-substring-search/zsh-history-substring-search.zsh" >> ~/.zshrc
